% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ESReport-package.R
\docType{package}
\name{ESReport-package}
\alias{ESReport-package}
\title{A Collection of Functions to Create Reports for Electron Spin Resonance
Dating Applications}
\description{
<placeholder>
}
\author{
Christoph Burow (University of Cologne, Germany)
}
\keyword{package}

