% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yamlHeader.R
\name{get_yamlHeader}
\alias{get_yamlHeader}
\title{YAML Header for ESR Reports}
\usage{
get_yamlHeader()
}
\value{
Console output.
}
\description{
Print YAML Header to be used for producing ESR reports.
}
\examples{

# Copy the output to your rmarkdown file
get_yamlHeader()

}

