% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_Report.R
\name{render_Report}
\alias{render_Report}
\title{Render R Markdown (Wrapper)}
\usage{
render_Report(file, output_format = c("pdf_document", "html_document"), ...)
}
\arguments{
\item{file}{\code{\link{character}} \bold{(required)}: Input file (R script, Rmd, or plain markdown).}

\item{output_format}{\code{\link{character}}:
R Markdown output format to convert to. Pass "all" to render all formats defined within the file.
Pass the name of a format (e.g. "html_document") to render a single format or pass a vector of
format names to render multiple formats. Alternatively you can pass an output format object; e.g. html_document().
If NULL is passed then the output format is the first one defined in the YAML
metadata of the input file (defaulting to HTML if none is specified).}

\item{...}{further arguments passed to \code{\link[rmarkdown]{render}}.}
}
\value{
A compiled document is written into the output file, and the path of the output file is returned.
}
\description{
Render the input file to the specified out ouput format using pandoc. This is
a convenient wrapper for \code{\link[rmarkdown]{render}} to produce HTML and
PDF output simultaneously.
}
\examples{

\dontrun{
render_Report("~/PATH/TO/FILE.rmd")
}

}
