% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_Comment.R
\name{report_Comment}
\alias{report_Comment}
\title{Insert a Comment Section}
\usage{
report_Comment(level = 1, text, delim)
}
\arguments{
\item{level}{\code{\link{integer}} \bold{(required)}: A non-negative integer
specifying the section level.}

\item{text}{\code{\link{character}} (optional): The comment text to be inserted
in the section.}

\item{delim}{\code{\link{character}} \bold{(required)}: Keyword specifying
the delimiter insterted before the section. Usually \code{"<br>"} or \code{"<hr>"}
for HTML and \code{"\\\\newpage"} for PDF reports.}
}
\value{
Text output via \code{\link{cat}}.
}
\description{
Insert a comment section in RMarkdown reports. Use this function within
a code chunk and make sure to set the chunk option \code{results = 'asis'}.
}
\examples{

report_Comment(level = 1, text = "My comment", delim = "\\\\newpage")

}

