% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_DE.R
\name{report_DE}
\alias{report_DE}
\title{Insert a Equivalent Dose Calculation Section}
\usage{
report_DE(files, natural = c(1, 2), dose, meta = NULL, settings = NULL,
  sample.weight, interval, model = "EXP", delim = "\\\\newpage")
}
\arguments{
\item{files}{\code{\link{character}} \bold{(required)}:
file path or directory where the spectra files are stored.}

\item{natural}{\code{\link{numeric}} (optional):}

\item{dose}{\code{\link{character}} or \code{\link{numeric}} \bold{(required)}:
Either a file path to a text file containing the irradiation doses or a numeric
vector.}

\item{meta}{\code{\link{list}} \bold{(required)}: A named \code{list} with
information on the following elements given in \code{\link{.meta}}.
Consider using \code{\link{.meta}} as a constructor.}

\item{settings}{\code{\link{list}} \bold{(required)}: A named \code{list} with
information on the following elements given in \code{\link{.settings}}.
Consider using \code{\link{.settings}} as a constructor.}

\item{sample.weight}{\code{\link{character}} or \code{\link{numeric}} (optional):
Either a file path to a text file containing the sample weights or a numeric
vector.}

\item{interval}{\code{\link{numeric}} \bold{(required)}:
A vector of length two specifying the range of x-values where peaks are searched}

\item{model}{\code{\link{character}} \bold{(required)}:
Currently implemented models: single-saturating exponential ("EXP"), linear ("LIN").}

\item{delim}{\code{\link{character}} \bold{(required)}: Keyword specifying
the delimiter insterted before the section. Usually \code{"<br>"} or \code{"<hr>"}
for HTML and \code{"\\\\newpage"} for PDF reports.}
}
\value{
Text output via \code{\link{cat}}.
}
\description{
Insert a section on the calculation of the equivalent dose for a given set
of input files. Use this function within a code chunk and make sure to set
the chunk option \code{results = 'asis'}.
}
\examples{

# not available yet

}

