% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_Helper.R
\name{report_Helper}
\alias{report_Helper}
\alias{.newPage}
\alias{.header}
\alias{.section}
\title{Helper functions for markdown sectioning}
\usage{
.newPage(delim)

.header(level, text)

.section(level, text, delim)
}
\arguments{
\item{delim}{\code{\link{character}} \bold{(required)}: Keyword specifying
the delimiter insterted before the section. Usually \code{"<br>"} or \code{"<hr>"}
for HTML and \code{"\\\\newpage"} for PDF reports.}

\item{level}{\code{\link{integer}} \bold{(required)}: A non-negative integer
specifying the section level.}

\item{text}{\code{\link{character}} (optional): The comment text to be inserted
in the section.}
}
\value{
Text output via \code{\link{cat}}.
}
\description{
A collection of helper functions for easier sectioning of markdown documents.
They take care that e.g. control commands for new lines (\code{"\\newline"} or \code{"<br>"})
are correctly positioned by adding \code{"\\n"} before and after a structure.
}
\examples{

.newPage("\\\\newpage")
.header(1, "My Header")
.section(2, "My Section", "<hr>")

}
