% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_Meta.R
\name{report_Meta}
\alias{.meta}
\alias{report_Meta}
\title{Insert a Meta-information Section}
\usage{
report_Meta(x = .meta())

.meta(person = NA, date = NA, file_suffix = NA, file_path = NA)
}
\arguments{
\item{x}{\code{\link{list}} \bold{(required)}: A named \code{list} with
information on the following elements given in \code{\link{.meta}}.
Consider using \code{\link{.meta}} as a constructor.}

\item{person}{\code{\link{character}} (optional): Name of the person that
conducted the experiment/measurement.}

\item{date}{\code{\link{character}} (optional): Date of measurement.}

\item{file_suffix}{\code{\link{character}} (optional): Do the raw measurement
files have a common file suffix (pattern)?}

\item{file_path}{\code{\link{character}} (optional): Path where the raw
measurement files are stored on the ESR spectrometer.}
}
\value{
Text output via \code{\link{cat}}.
}
\description{
Insert meta information in RMarkdown reports. Use this function within
a code chunk and make sure to set the chunk option \code{results = 'asis'}.
}
\examples{

report_Meta(x = list("Max Mustermann", "Yesterday", c("_a", "_b"), "~/ESR/"))

report_Meta(.meta(person = "Max Mustermann",
                  date = "Yesterday",
                  file_suffix = c("_a", "_b"),
                  file_path = "~/ESR/"))

}

