% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_Sample.R
\name{report_Sample}
\alias{report_Sample}
\alias{.sample}
\title{Insert a Sample Information Section}
\usage{
report_Sample(x = .sample(), delim)

.sample(sample = NA, type = NA, exp_age = NA, n_aliquots = NA,
  measurement_dates = NA, irr_doses = NA)
}
\arguments{
\item{x}{\code{\link{list}} \bold{(required)}: A named \code{list} with
information on the following elements given in \code{\link{.sample}}.
Consider using \code{\link{.sample}} as a constructor.}

\item{delim}{\code{\link{character}} \bold{(required)}: Keyword specifying
the delimiter insterted before the section. Usually \code{"<br>"} or \code{"<hr>"}
for HTML and \code{"\\\\newpage"} for PDF reports.}

\item{sample}{\code{\link{character}}: Sample name or laboratory code.}

\item{type}{\code{\link{character}}: Sample type (e.g., "Coral" or "Mollusc").}

\item{exp_age}{\code{\link{character}}: The expected age range.}

\item{n_aliquots}{\code{\link{character}}: Number of aliquots used to determine
the equivalent dose.}

\item{measurement_dates}{\code{\link{character}}: When were the measurements
conducted?}

\item{irr_doses}{\code{\link{character}}: Irradiation doses (in Gy) of the
individual aliquots.}
}
\value{
Text output via \code{\link{cat}}.
}
\description{
Insert a sample information section in RMarkdown reports. Use this function within
a code chunk and make sure to set the chunk option \code{results = 'asis'}.
}
\examples{

report_Sample(x = list("K9999", "Coral", "Holocene", 20, "29/02/2016", 1:20),
              delim = "\\\\newpage")

report_Sample(.sample(sample = "K9999",
                      type = "Coral",
                      exp_age = "Holocene",
                      n_aliquots = 20,
                      measurement_dates = "29/02/2016",
                      irr_dose = 1:20),
               delim = "\\\\newpage")

}
