% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_Screening.R
\name{report_Screening}
\alias{report_Screening}
\title{Insert a Screening Section}
\usage{
report_Screening(files, meta = NULL, settings = NULL, delim)
}
\arguments{
\item{files}{\code{\link{character}} \bold{(required)}: file path or directory where the spectra files are stored.}

\item{meta}{\code{\link{list}} \bold{(required)}: A named \code{list} with
information on the following elements given in \code{\link{.meta}}.
Consider using \code{\link{.meta}} as a constructor.}

\item{settings}{\code{\link{list}} \bold{(required)}: A named \code{list} with
information on the following elements given in \code{\link{.settings}}.
Consider using \code{\link{.settings}} as a constructor.}

\item{delim}{\code{\link{character}} \bold{(required)}: Keyword specifying
the delimiter insterted before the section. Usually \code{"<br>"} or \code{"<hr>"}
for HTML and \code{"\\\\newpage"} for PDF reports.}
}
\value{
Text output via \code{\link{cat}}.
}
\description{
Insert a spectrum screening section in RMarkdown reports. Use this function within
a code chunk and make sure to set the chunk option \code{results = 'asis'}.
}
\examples{

\dontrun{
report_Screening(files = "~/PATH/TO/FILES/",
                 meta = .meta(person = "Max Mustermann",
                              date = "Yesterday",
                              file_suffix = c("_a", "_b"),
                              file_path = "~/ESR/"),
                 settings = .settings(device = "Bruker ESP300-E",
                                      mod_amp = 0.485,
                                      receiver_gain = 2,
                                      freq = 9.628,
                                      conv_time = 20.48,
                                      time_const = 163.84,
                                      sweep_time = 20.972,
                                      power = 25.3,
                                      center_field = 3430,
                                      sweep_width = 40,
                                      scans = c(10, 20, 30, 40)),
                 delim = "\\\\newpage")
}

}
