% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_Settings.R
\name{report_Settings}
\alias{.settings}
\alias{report_Settings}
\title{Insert a Experimental Settings Section}
\usage{
report_Settings(x = .settings())

.settings(device = c("Bruker ESP300-E", "Bruker ELEXSYS 500")[1],
  mod_amp = NA, receiver_gain = NA, freq = NA, conv_time = NA,
  time_const = NA, sweep_time = NA, power = NA, center_field = NA,
  sweep_width = NA, scans = NA, file = NA)
}
\arguments{
\item{x}{\code{\link{list}} \bold{(required)}: A named \code{list} with
information on the following elements given in \code{\link{.settings}}.
Consider using \code{\link{.settings}} as a constructor.}

\item{device}{\code{\link{character}}: Name or model number of the ESR spectrometer.}

\item{mod_amp}{\code{\link{numeric}}: Modulation Amplitude}

\item{receiver_gain}{\code{\link{numeric}}: Receiver Gain}

\item{freq}{\code{\link{numeric}}: Microwave Frequency}

\item{conv_time}{\code{\link{numeric}}: Conversion Time}

\item{time_const}{\code{\link{numeric}}: Time Constant}

\item{sweep_time}{\code{\link{numeric}}: Sweep Time}

\item{power}{\code{\link{numeric}}: Microwave Power}

\item{center_field}{\code{\link{integer}}: Center Field}

\item{sweep_width}{\code{\link{integer}}: Sweep Width}

\item{scans}{\code{\link{integer}}: Number of scans}

\item{file}{\code{\link{character}}: File name}
}
\value{
Text output via \code{\link{cat}}.
}
\description{
Insert a section on spectrometer settings in RMarkdown reports. Use this function within
a code chunk and make sure to set the chunk option \code{results = 'asis'}.
}
\examples{

report_Settings(.settings(device = "Bruker ESP300-E",
                          mod_amp = 0.485,
                          receiver_gain = 2,
                          freq = 9.628,
                          conv_time = 20.48,
                          time_const = 163.84,
                          sweep_time = 20.972,
                          power = 25.3,
                          center_field = 3430,
                          sweep_width = 40,
                          scans = c(10, 20, 30, 40)))

}

