% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_DE.R
\name{report_DE}
\alias{report_DE}
\title{Insert a Equivalent Dose Calculation Section}
\usage{
report_DE(files, title = "Equivalent dose estimation", title.suffix = "",
  natural = c(1, 2), natural.comment, dose, meta = NULL, settings = NULL,
  fit.repeated = FALSE, sample.weight = NULL, interval, th = 10,
  model = "EXP", amplitudes = NULL, spike = NULL, bootstrap = TRUE,
  delim = "\\\\newpage", ...)
}
\arguments{
\item{files}{\code{\link{character}} \bold{(required)}:
file path or directory where the spectra files are stored.}

\item{title}{\code{\link{character}} (optional):
Main title of the section (defaults to 'Equivalent dose estimation').}

\item{title.suffix}{\code{\link{character}} (optional):
Optional text to appear in brackets after the title. Useful if the report
contains multiple DE measurements.}

\item{natural}{\code{\link{numeric}} (optional):
Optional vector of integer values specifying which of the spectra are
repeated natural dose measurements (defaults to \code{c(1, 2)}).}

\item{natural.comment}{\code{\link{character}} (optional):
An optional comment on the natural dose. Defaults to the statement that
repeated natural dose measurements were done first and last in the
measurement series.}

\item{dose}{\code{\link{character}} or \code{\link{numeric}} \bold{(required)}:
Either a file path to a text file containing the irradiation doses or a numeric
vector.}

\item{meta}{\code{\link{list}} \bold{(required)}: A named \code{list} with
information on the following elements given in \code{\link{.meta}}.
Consider using \code{\link{.meta}} as a constructor.}

\item{settings}{\code{\link{list}} \bold{(required)}: A named \code{list} with
information on the following elements given in \code{\link{.settings}}.
Consider using \code{\link{.settings}} as a constructor.}

\item{fit.repeated}{\code{\link{logical}} (optional):
Whether a repeated zero-dose measurement should be considered
for fitting or not. If `TRUE` all repeated zero-dose measurements (indicated by
argument `natural`) are removed from the data set before fitting.}

\item{sample.weight}{\code{\link{character}} or \code{\link{numeric}} (optional):
Either a file path to a text file containing the sample weights or a numeric
vector.}

\item{interval}{\code{\link{numeric}} \bold{(required)}:
A vector of length two specifying the range of x-values where peaks are searched}

\item{th}{\code{\link{numeric}} (optional):
An integer specifying the number of neighbouring values to compare each x-value to.}

\item{model}{\code{\link{character}} \bold{(required)}:
Currently implemented models: single-saturating exponential ("EXP"), linear ("LIN").}

\item{amplitudes}{\code{\link{numeric}} (optional):
A numeric vector of signal amplitudes. If provided, automatic peak finding and
amplitude calculation is skipped.}

\item{spike}{\code{\link{character}} (optional):
Path to a single file of a spike ESR spectrum. The amplitude of the spike
signal will be subtracted from all other amplitudes (default \code{TRUE}).}

\item{bootstrap}{\code{\link{logical}} (optional):
Include calculation of the equivalent dose using a bootstrap approach.}

\item{delim}{\code{\link{character}} \bold{(required)}: Keyword specifying
the delimiter insterted before the section. Usually \code{"<br>"} or \code{"<hr>"}
for HTML and \code{"\\\\newpage"} for PDF reports.}

\item{...}{further arguments passed to \code{\link[ESR]{fit_DRC}},
\code{\link[ESR]{calc_DePlateau}} and \code{\link[ESR]{read_Spectrum}}.}
}
\value{
Text output via \code{\link{cat}}.
}
\description{
Insert a section on the calculation of the equivalent dose for a given set
of input files. Use this function within a code chunk and make sure to set
the chunk option \code{results = 'asis'}.
}
\examples{

# not available yet

}
